requirements.txt
lightgbm
google-generativeai
python-dotenv
PyMuPDF
xgboost